modified JUnit5 API to be more suitable for working with both @TestFactory and @TestTemplate methods
